Move get schema functions, fix inconsistent docstring, add pydantic to pyproject
