+++ camel/functions/__init__.py
@@ -12,14 +12,20 @@
 # limitations under the License.
 # =========== Copyright 2023 @ CAMEL-AI.org. All Rights Reserved. ===========
 
+from .openai_function import (
+    OpenAIFunction,
+    get_openai_tool_schema,
+    get_openai_function_schema,
+)
 from .math_functions import MATH_FUNCS
-from .openai_function import OpenAIFunction
 from .search_functions import SEARCH_FUNCS
 from .weather_functions import WEATHER_FUNCS
 from .unstructured_io_fuctions import UnstructuredModules
 
 __all__ = [
     'OpenAIFunction',
+    'get_openai_tool_schema',
+    'get_openai_function_schema',
     'MATH_FUNCS',
     'SEARCH_FUNCS',
     'WEATHER_FUNCS',
