+++ camel/utils/__init__.py
@@ -20,8 +20,6 @@ from .commons import (
     download_tasks,
     get_task_list,
     check_server_running,
-    get_openai_tool_schema,
-    get_openai_function_schema,
 )
 from .token_counting import (
     get_model_encoding,
@@ -43,6 +41,4 @@ __all__ = [
     'BaseTokenCounter',
     'OpenAITokenCounter',
     'OpenSourceTokenCounter',
-    'get_openai_tool_schema',
-    'get_openai_function_schema',
 ]
