+++ camel/utils/commons.py
@@ -17,26 +17,10 @@ import socket
 import time
 import zipfile
 from functools import wraps
-from inspect import Parameter, signature
-from typing import (
-    Any,
-    Callable,
-    Dict,
-    List,
-    Mapping,
-    Optional,
-    Set,
-    Tuple,
-    TypeVar,
-    cast,
-)
+from typing import Any, Callable, List, Optional, Set, TypeVar, cast
 from urllib.parse import urlparse
 
 import requests
-from docstring_parser import parse
-from pydantic import create_model
-from pydantic.alias_generators import to_pascal
-from pydantic.fields import FieldInfo
 
 from camel.types import TaskType
 
@@ -194,119 +178,3 @@ def check_server_running(server_url: str) -> bool:
 
     # if the port is open, the result should be 0.
     return result == 0
-
-
-def _remove_a_key(d: Dict, remove_key: Any) -> None:
-    r"""Remove a key from a dictionary recursively."""
-    if isinstance(d, dict):
-        for key in list(d.keys()):
-            if key == remove_key:
-                del d[key]
-            else:
-                _remove_a_key(d[key], remove_key)
-
-
-def get_openai_function_schema(func: Callable) -> Dict[str, Any]:
-    r"""Generates a schema dict for an OpenAI function based on its signature.
-
-    This function is deprecated and will be replaced by
-    :obj:`get_openai_tool_schema()` in future versions. It parses the
-    function's parameters and docstring to construct a JSON schema-like
-    dictionary.
-
-    Args:
-        func (Callable): The OpenAI function to generate the schema for.
-
-    Returns:
-        Dict[str, Any]: A dictionary representing the JSON schema of the
-            function, including its name, description, and parameter
-            specifications.
-    """
-    openai_function_schema = get_openai_tool_schema(func)["function"]
-    return openai_function_schema
-
-
-def get_openai_tool_schema(func: Callable) -> Dict[str, Any]:
-    r"""Generates an OpenAI JSON schema from a given Python function.
-
-    This function creates a schema compatible with OpenAI's API specifications,
-    based on the provided Python function. It processes the function's
-    parameters, types, and docstrings, and constructs a schema accordingly.
-
-    Note:
-        - Each parameter in `func` must have a type annotation; otherwise, it's
-          treated as 'Any'.
-        - Variable arguments (*args) and keyword arguments (**kwargs) are not
-          supported and will be ignored.
-        - A functional description including a brief and detailed explanation
-          should be provided in the docstring of `func`.
-        - All parameters of `func` must be described in its docstring.
-        - Supported docstring styles: ReST, Google, Numpydoc, and Epydoc.
-
-    Args:
-        func (Callable): The Python function to be converted into an OpenAI
-                         JSON schema.
-
-    Returns:
-        Dict[str, Any]: A dictionary representing the OpenAI JSON schema of
-                        the provided function.
-
-    See Also:
-        `OpenAI API Reference
-            <https://platform.openai.com/docs/api-reference/assistants/object>`_
-    """
-    params: Mapping[str, Parameter] = signature(func).parameters
-    fields: Dict[str, Tuple[type, FieldInfo]] = {}
-    for param_name, p in params.items():
-        param_type = p.annotation
-        param_default = p.default
-        param_kind = p.kind
-        param_annotation = p.annotation
-        # Variable parameters are not supported
-        if (param_kind == Parameter.VAR_POSITIONAL
-                or param_kind == Parameter.VAR_KEYWORD):
-            continue
-        # If the parameter type is not specified, it defaults to typing.Any
-        if param_annotation is Parameter.empty:
-            param_type = Any
-        # Check if the parameter has a default value
-        if param_default is Parameter.empty:
-            fields[param_name] = (param_type, FieldInfo())
-        else:
-            fields[param_name] = (param_type, FieldInfo(default=param_default))
-
-    # Applying `create_model()` directly will result in a mypy error,
-    # create an alias to avoid this.
-    def _create_mol(name, field):
-        return create_model(name, **field)
-
-    model = _create_mol(to_pascal(func.__name__), fields)
-    parameters_dict = model.model_json_schema()
-    # The "title" is generated by model.model_json_schema()
-    # but is useless for openai json schema
-    _remove_a_key(parameters_dict, "title")
-
-    docstring = parse(func.__doc__ or "")
-    for param in docstring.params:
-        if ((name := param.arg_name) in parameters_dict["properties"]
-                and (description := param.description)):
-            parameters_dict["properties"][name]["description"] = description
-
-    short_description = docstring.short_description or ""
-    long_description = docstring.long_description or ""
-    if long_description:
-        func_description = f"{short_description}\n{long_description}"
-    else:
-        func_description = short_description
-
-    openai_function_schema = {
-        "name": func.__name__,
-        "description": func_description,
-        "parameters": parameters_dict,
-    }
-
-    openai_tool_schema = {
-        "type": "function",
-        "function": openai_function_schema
-    }
-    return openai_tool_schema
