+++ pyproject.toml
@@ -36,6 +36,7 @@ jsonschema = "^4"
 protobuf = "^4"
 pathlib = "^1.0.1"   
 docstring-parser = "^0.15"
+pydantic = ">=1,<3"
 
 # huggingface-agent
 transformers = { version = "^4", optional = true }
